[hammerspoon] Adds hammerspoon configuration
AeroSpace is nice and all, but cumbersome to use on an UW screen.
Because I'm used to certain apps living in specific workspaces and the
keybinds to get there from AeroSpace I miss them when AeroSpace is
disabled.

Hammerspoon comes in clutch here as it lets me bind keys to apps